config.py
### File: config.py
###SECRET_KEY = 'a1f8fc3e0a4196acd75918ef69da5b961751a8cdcaac060c'
##MONGO_URI = 'mongodb://localhost:27017/petcare'
import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

class Config:
    # Flask configuration
    SECRET_KEY = os.getenv('a1f8fc3e0a4196acd75918ef69da5b961751a8cdcaac060c', 'dev-secret-key')
    DEBUG = os.getenv('FLASK_DEBUG', 'False').lower() == 'true'
    
    # MongoDB configuration
    MONGO_URI = os.getenv('MONGO_URI', 'mongodb://localhost:27017/petcare')
    
    # Application configuration
    APP_NAME = 'PetCare Management System'
    ADMIN_EMAIL = os.getenv('ADMIN_EMAIL', 'admin@example.com')
    
    # File upload configuration
    UPLOAD_FOLDER = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'static/uploads')
    ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif'}
    MAX_CONTENT_LENGTH = 16 * 1024 * 1024  # 16MB max upload size