requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Flask-WTF==1.0.0
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
pymongo==4.0.1
python-dotenv==0.19.1
email-validator==1.1.3